+++ src/main/java/org/takes/facets/fallback/TkFallback.java
@@ -34,8 +34,8 @@ import org.takes.Request;
 import org.takes.Response;
 import org.takes.Take;
 import org.takes.misc.Opt;
+import org.takes.rq.RqBaseMethod;
 import org.takes.rq.RqHref;
-import org.takes.rq.RqMethod;
 import org.takes.rs.ResponseOf;
 import org.takes.tk.TkWrap;
 
@@ -231,7 +231,7 @@ public final class TkFallback extends TkWrap {
         return new IllegalStateException(
             String.format(
                 "[%s %s] failed in %s: %s",
-                new RqMethod.Base(req).method(),
+                new RqBaseMethod(req).method(),
                 new RqHref.Base(req).href(),
                 time, TkFallback.msg(exp)
             ),
