+++ src/main/java/org/takes/rq/RqMethod.java
@@ -24,9 +24,6 @@
 package org.takes.rq;
 
 import java.io.IOException;
-import java.util.Locale;
-import java.util.regex.Pattern;
-import lombok.EqualsAndHashCode;
 import org.takes.Request;
 
 /**
@@ -89,42 +86,4 @@ public interface RqMethod extends Request {
      * @throws IOException If fails
      */
     String method() throws IOException;
-
-    /**
-     * Request decorator, for HTTP method parsing.
-     *
-     * <p>The class is immutable and thread-safe.
-     *
-     * @since 0.13.7
-     */
-    @EqualsAndHashCode(callSuper = true)
-    final class Base extends RqWrap implements RqMethod {
-
-        /**
-         * HTTP token separators which are not already excluded by PATTERN.
-         */
-        private static final Pattern SEPARATORS = Pattern.compile(
-            "[()<>@,;:\\\"/\\[\\]?={}]"
-        );
-
-        /**
-         * Ctor.
-         * @param req Original request
-         */
-        public Base(final Request req) {
-            super(req);
-        }
-
-        @Override
-        public String method() throws IOException {
-            final String method = new RqRequestLine.Base(this)
-                .method();
-            if (Base.SEPARATORS.matcher(method).find()) {
-                throw new IOException(
-                    String.format("Invalid HTTP method: %s", method)
-                );
-            }
-            return method.toUpperCase(Locale.ENGLISH);
-        }
-    }
 }
